train2/requirements.txt
Django==1.11.21
ipython==4.1.2
django-extensions==1.7.5
djangorestframework==3.5.3
pytz==2018.9
xlrd==1.0.0
psycopg2-binary==2.8.3
requests==2.12.5
django-rest-swagger==2.1.1
django-cors-headers==2.0.0
numpy==1.16.0
pandas==0.23.4
geocoder==1.22.6
Fabric3==1.13.1.post1
tabulate==0.8.1
sphinx_rtd_theme==0.4.0
pymessenger==0.0.7.0
partridge==1.0.0
python-Levenshtein==0.12.0
python-telegram-bot==11.1.0
